// src/enums/meleeWeaponName.ts
export enum MeleeWeaponName {
	LanceOfIlluminationStrike = 'Lance of Illumination – strike',
	LanceOfIlluminationSweep = 'Lance of Illumination – sweep',
	BlessedBlade = 'Blessed blade',
	Chainsword = 'Chainsword',
	PowerWeapon = 'Power weapon',
	PalatineBlade = 'Palatine blade',
	MaceOfCastigation = 'Mace of Castigation',
	TheArdentBlade = 'The Ardent Blade',
	RelicWeapons = 'Relic weapons',
	Sanctity = 'Sanctity',
	TheOutcastsWeapons = 'The Outcast’s Weapons',
	CloseCombatWeapon = 'Close combat weapon',
	NovitiateMeleeWeapons = 'Novitiate melee weapons',
	BladeOfVigil = 'Blade of Vigil',
	ScribesStaff = 'Scribe’s staff',
	DialogusStaff = 'Dialogus staff',
	ZealotsVindictor = 'Zealot’s vindictor',
	AnointedHalberd = 'Anointed halberd',
	HallowedMace = 'Hallowed mace',
	SpearOfTheFaithful = 'Spear of the faithful',
	ChirurgeonsTools = 'Chirurgeon’s tools',
	MaceOfTheRighteous = 'Mace of the righteous',
	ParagonWarBlade = 'Paragon war blade',
	ParagonWarMace = 'Paragon war mace',
	NeuralWhips = 'Neural whips',
	PenitentEviscerator = 'Penitent eviscerator',
	DeathCultPowerBlades = 'Death Cult power blades',
	PenitentBuzzBlade = 'Penitent buzz-blade',
	PenitentFlail = 'Penitent flail',
	TwinPenitentBuzzBlades = 'Twin penitent buzz-blades',
	TwinPenitentFlails = 'Twin penitent flails',
	ArmouredTracks = 'Armoured tracks',
	ArcoFlails = 'Arco-flails'
}