// src/enums/unitName.ts
export enum UnitName {
	MorvennVahl = 'Morvenn Vahl',
	Canoness = 'Canoness',
	Palatine = 'Palatine',
	JunithEruita = 'Junith Eruita',
	Missionary = 'Missionary',
	SaintCelestine = 'Saint Celestine',
	GeminaeSuperia = 'Geminae Superia',
	TriumphOfSaintKatherine = 'Triumph of Saint Katherine',
	EphraelStern = 'Ephrael Stern',
	KyganilOfTheBloodyTears = 'Kyganil of the Bloody Tears',
	BattleSisterSquad = 'Battle Sister Squad',
	NovitiateSuperior = 'Novitiate Superior',
	SisterNovitiate = 'Sister Novitiate',
	AestredThurga = 'Aestred Thurga',
	AgathaeDolan = 'Agathae Dolan',
	Imagifier = 'Imagifier',
	Dialogus = 'Dialogus',
	Preacher = 'Preacher',
	CelestianSacresants = 'Celestian Sacresants',
	Hospitaller = 'Hospitaller',
	Dogmata = 'Dogmata',
	ParagonWarsuits = 'Paragon Warsuits',
	RepentiaSuperior = 'Repentia Superior',
	SistersRepentia = 'Sisters Repentia',
	Crusaders = 'Crusaders',
	DeathCultAssassins = 'Death Cult Assassins',
	DominionSquad = 'Dominion Squad',
	SeraphimSquad = 'Seraphim Squad',
	ZephyrimSquad = 'Zephyrim Squad',
	RetributorSquad = 'Retributor Squad',
	Mortifiers = 'Mortifiers',
	PenitentEngines = 'Penitent Engines',
	Exorcist = 'Exorcist',
	Castigator = 'Castigator',
	SororitasRhino = 'Sororitas Rhino',
	Immolator = 'Immolator',
	ArcoFlagellants = 'Arco-Flagellants'
}